refactor(embedding): decompose into focused files and smaller functions
Split embedding.go (849 lines, 5 mixed concerns, RunFromDB CC ~22) into
5 single-responsibility files:

- embed_errors.go: sentinel errors, EmbedHTTPError, deferredJob
- ollama.go: OllamaClient, HTTP transport, request/response types
- circuit_breaker.go: CircuitBreaker state machine
- embed_worker.go: EmbedWorker, RunFromDB orchestrator, helpers
- embed_cache.go: LRU query embedding cache

RunFromDB decomposed from 278 lines (CC ~22) into 4 focused functions:
- RunFromDB: slim 45-line orchestrator (Phase 1 pagination + Phase 2 call)
- reconcileAndEmbed: Has() check → embed → upsert → mark per batch
- embedWithRetry: 3-quick-retry loop with CB and adaptive batch sizing
- retryDeferred: Phase 2 multi-pass single-chunk retry

Introduced runState struct to group mutable progress (embedded, skipped,
activeBatch, deferred, total, onProgress) instead of 6+ local variables.

No behavior changes. All existing tests pass without modification.
